refactor: move type
Move env type into corresponding module
